+++ sentry-android-core/src/test/java/io/sentry/android/core/PermissionsTest.kt
@@ -4,7 +4,7 @@ import android.Manifest
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.sentry.android.core.util.Permissions
+import io.sentry.android.core.internal.util.Permissions
 import org.junit.runner.RunWith
 import kotlin.test.BeforeTest
 import kotlin.test.Test