+++ sentry-android-core/src/test/java/io/sentry/android/core/internal/util/DeviceOrientationsTest.kt
@@ -1,10 +1,10 @@
-package io.sentry.android.core.util
+package io.sentry.android.core.internal.util
 
 import android.content.res.Configuration.ORIENTATION_LANDSCAPE
 import android.content.res.Configuration.ORIENTATION_PORTRAIT
 import android.content.res.Configuration.ORIENTATION_SQUARE
 import android.content.res.Configuration.ORIENTATION_UNDEFINED
-import io.sentry.android.core.util.DeviceOrientations.getOrientation
+import io.sentry.android.core.internal.util.DeviceOrientations.getOrientation
 import io.sentry.protocol.Device
 import kotlin.test.Test
 import kotlin.test.assertEquals